+++ RUFAS/routines/animal/manure/dry_cow_manure_excretion.py
@@ -5,6 +5,7 @@ from RUFAS.data_structures.animal_manure_excretions import AnimalManureExcretion
 from RUFAS.general_constants import GeneralConstants
 from RUFAS.routines.animal.animal_module_constants import AnimalModuleConstants
 from RUFAS.routines.animal.manure.general_manure import calculate_phosphorus_excretion_values
+from RUFAS.user_constants import UserConstants
 
 
 def manure_calculations(
@@ -125,7 +126,7 @@ def manure_calculations(
         15.1
         + 0.83
         * (dry_matter_intake * GeneralConstants.KG_TO_GRAMS)
-        * (CP_concentration * GeneralConstants.PROTEIN_TO_NITROGEN)
+        * (CP_concentration * UserConstants.PROTEIN_TO_NITROGEN)
         / GeneralConstants.FRACTION_TO_PERCENTAGE
     ) * GeneralConstants.GRAMS_TO_KG
 
@@ -134,7 +135,7 @@ def manure_calculations(
         0.345
         + 0.317
         * (dry_matter_intake * GeneralConstants.KG_TO_GRAMS)
-        * (CP_concentration * GeneralConstants.PROTEIN_TO_NITROGEN)
+        * (CP_concentration * UserConstants.PROTEIN_TO_NITROGEN)
         / GeneralConstants.FRACTION_TO_PERCENTAGE
     ) * GeneralConstants.GRAMS_TO_KG